+++ frontend/src/App.jsx
@@ -1,17 +1,19 @@
 import { Routes, Route } from "react-router-dom";
-import Home from "./Pages/Home";
+import Home from "./pages/Home";
 import "./index.css";
-import EnregistrerTel1 from "./Pages/EnregistrerTel1";
-import EnregistrerTel2 from "./Pages/EnregistrerTel2";
-import EnregistrerTel3 from "./Pages/EnregistrerTel3";
-import ConnectionPage from "./Pages/ConnectionPage";
-import ResumeTel from "./Pages/ResumeTel";
+import EnregistrerTel1 from "./pages/EnregistrerTel1";
+import EnregistrerTel2 from "./pages/EnregistrerTel2";
+import EnregistrerTel3 from "./pages/EnregistrerTel3";
+import ConnectionPage from "./pages/ConnectionPage";
+import ResumeTel from "./pages/ResumeTel";
+import TableauSmartphones from "./pages/TableauSmartphones";
 
 function App() {
   return (
     <Routes>
       <Route path="/" element={<ConnectionPage />} />
-      <Route path="/home" element={<Home />} />
+      <Route path="/Home" element={<Home />} />
+      <Route path="/TableauSmartphones" element={<TableauSmartphones />} />
       <Route path="/enregistrer1" element={<EnregistrerTel1 />} />
       <Route path="/enregistrer2" element={<EnregistrerTel2 />} />
       <Route path="/enregistrer3" element={<EnregistrerTel3 />} />
