+++ frontend/tailwind.config.js
@@ -9,7 +9,6 @@ module.exports = {
     },
     colors: {
       grey: "#F2F2F2",
-
       primary: "#002743",
       secondary: "#E62460",
       tertiary: "#00ACB0",
